requirements.txt
langchain
langchain-openai
langgraph
pydantic
python-dotenv
requests
beautifulsoup4
pandas
lxml
langsmith
openevals
datasets
openai
